Re-arrange common module
